I trace 22 find precison go down reason in maven

+++ src/cn/edu/nju/cs/itrace4/exp/infinispan/preprocess/PreprocessTextInfinispan.java
@@ -1,14 +1,17 @@
 package cn.edu.nju.cs.itrace4.exp.infinispan.preprocess;
 
+
 import java.io.File;
 
 import cn.edu.nju.cs.itrace4.core.type.Granularity;
 import cn.edu.nju.cs.itrace4.exp.tool.GetSrc;
+import cn.edu.nju.cs.itrace4.exp.tool.GetSrcBaseRI;
 import cn.edu.nju.cs.itrace4.exp.tool.GetUC;
 import cn.edu.nju.cs.itrace4.exp.tool.TransferTXT;
 import cn.edu.nju.cs.itrace4.parser.SourceTargetUnionForGit;
 import cn.edu.nju.cs.itrace4.preprocess.BatchingPreprocess;
 import cn.edu.nju.cs.itrace4.preprocess.rawdata.db.GenerateRTM;
+import cn.edu.nju.cs.itrace4.preprocess.rawdata.db.GenerateRTMThroughCluster;
 import cn.edu.nju.cs.itrace4.relation.RelationInfo;
 
 /**
@@ -19,11 +22,14 @@ import cn.edu.nju.cs.itrace4.relation.RelationInfo;
 public class PreprocessTextInfinispan {
 	private GenerateRTM getRTM;
 	private GetUC getUC = new GetUC();
-	private GetSrc getOriginSrc = new GetSrc();
+	
+	//private GetSrc getOriginSrc = new GetSrc(rtmDBFilePath);
+	private GetSrc getOriginSrc = new GetSrcBaseRI(relationDirPath);
+	
 	private TransferTXT getSrc = new TransferTXT();
 	//private TableFormatNormalize generateCallGraph = new TableFormatNormalize();
 	
-	
+	private String clusterFilePath = "data/exp/Infinispan/clusterFile/s_d_t_120d.txt";
 	private static String projectPath = "data/exp/Infinispan/";
 
     private static String rtmDBFilePath = projectPath + "rtm/Infinispan-req.db";
@@ -38,12 +44,12 @@ public class PreprocessTextInfinispan {
     private static String masterPath = projectPath+"infinispan-master";
     private static String graphDBPath = relationDirPath + File.separator + "call.db";
     
-    private String dbProperty = "resource/infinispanDB.property";
+    private String dbProperty = "resource/InfinispanDB.property";
     private String sqlFile = "resource/sql/buildRTMForInfinispan.sql";
     
-    
     public PreprocessTextInfinispan() {
-    	getRTM = new GenerateRTM(rtmDBFilePath, dbProperty, sqlFile);
+    	//getRTM = new GenerateRTM(rtmDBFilePath,dbProperty,sqlFile);
+    	getRTM = new GenerateRTMThroughCluster(rtmDBFilePath,dbProperty,sqlFile,clusterFilePath);
     }
     
     private void cleanData() {
@@ -55,8 +61,13 @@ public class PreprocessTextInfinispan {
     private void deleteFileInDir(String dirPath) {
     	File dir = new File(dirPath);
 		File[] files = dir.listFiles();
-		for(File f:files) {
-			f.delete();
+		if(files==null) {
+			return ;
+		}
+		else {
+			for(File f:files) {
+				f.delete();
+			}
 		}
 	}
 
@@ -67,7 +78,11 @@ public class PreprocessTextInfinispan {
 			getUC.getUCFromDB(ucDirPath,rtmDBFilePath);
 			String originPath = srcDirPath;
 	    	String targetPath = classDirPath;
-	    	getOriginSrc.getSrcFromMasterBasedOnGraphDB(masterPath, srcDirPath, graphDBPath);
+	    	//getOriginSrc.getSrcFromMasterBasedOnGraphDB(masterPath, srcDirPath, graphDBPath);
+//	    	getOriginSrc.getSrcFromMasterBasedOnGraphDBNewFormatDB(masterPath,originPath,graphDBPath,"callGraph",
+//	    			"caller","callee");
+	    	getOriginSrc.getSrcBaseRI(masterPath, originPath);
+	    	
 	    	getSrc.transferTXT(originPath, targetPath);
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -87,15 +102,14 @@ public class PreprocessTextInfinispan {
          *              step5:  get union between class,uc,ri and rtm.
          *              step6:  process txt. 
          */
-    	PreprocessTextInfinispan infinispanProcess = new PreprocessTextInfinispan();
-    	infinispanProcess.arrangeData();
+    	PreprocessTextInfinispan InfinispanProcess = new PreprocessTextInfinispan();
+    	InfinispanProcess.arrangeData();
     	
     	SourceTargetUnionForGit union = new SourceTargetUnionForGit(ucDirPath, srcDirPath, rtmDBFilePath, Granularity.CLASS,classDirPath,methodDirPath);
 
         BatchingPreprocess preprocess = new BatchingPreprocess(ucDirPath, classDirPath, methodDirPath);
         preprocess.doProcess();
         
-        //System.setProperty("projectType", "git");
         RelationInfo rg = new RelationInfo(classDirPath, relationDirPath, Granularity.CLASS);
         rg.showMessage();
     }
@@ -106,9 +120,9 @@ public class PreprocessTextInfinispan {
 //         * @date 2017/10/11
 //         * @description:为了搞清楚<code>预处理</code>和 <code>序列化</code>的整体流程，先把无关的注释掉
 //         */
-//    	PreprocessTextInfinispan infinispanProcess = new PreprocessTextInfinispan(rtmDBFilePath);
-//    	infinispanProcess.cleanData();
-//    	infinispanProcess.arrangeData();
+//    	PreprocessTextInfinispan InfinispanProcess = new PreprocessTextInfinispan(rtmDBFilePath);
+//    	InfinispanProcess.cleanData();
+//    	InfinispanProcess.arrangeData();
 //    	
 //    	SourceTargetUnionForGit union = new SourceTargetUnionForGit(ucDirPath, srcDirPath, rtmDBFilePath, Granularity.CLASS,classDirPath,methodDirPath);
 //
